# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Animal Poke Backend API
  version: 1.0.0
  description: |
    Go 后端 API 契约（OpenAPI 3.1）。
    客户端仅持有 API Base URL + 设备 Token；第三方 Key 永不进入前端。
  contact:
    name: animal_poke team
  license:
    name: Proprietary
    url: https://github.com/Bluepoisons/animal_poke
servers:
  - url: https://api.animal-poke.app
    description: Production
  - url: https://api.staging.animal-poke.app
    description: Staging
...
paths:
...
  /api/v1/sync/animals/{uuid}:
    get:
      tags: [Sync]
      summary: Get single collection animal detail
      description: |
        Returns one owned animal by uuid. Soft-deleted animals return a minimal
        tombstone (`uuid`, `deleted_at`, `server_version`). Unauthorized access
        returns 404 (no existence leak).
      operationId: getAnimalDetail
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Animal detail or tombstone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionAnimalDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Not found or not owned
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    patch:
      tags: [Sync]
      summary: Patch collection animal (nickname/favorite/locked)
      description: |
        Partial update of client-editable collection metadata.
        Optimistic lock via `If-Match` header (preferred) or body `server_version`.
        On version conflict returns 409 with `current` representation for merge.
      operationId: patchAnimal
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: false
          schema:
            type: string
          description: Expected server_version (quoted ETag supported)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionAnimalPatch"
      responses:
        "200":
          description: Updated animal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionAnimalDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Not found or not owned
        "409":
          description: Version conflict / already deleted / locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionVersionConflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags: [Sync]
      summary: Soft-delete animal (tombstone)
      description: |
        Soft-deletes one animal and bumps server_version for pull.
        Response is a minimal tombstone only — never full content.
        Requires If-Match or query server_version. Locked animals return 409.
        Double delete returns 409 already_deleted with current tombstone.
      operationId: deleteAnimal
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: false
          schema:
            type: string
        - name: server_version
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Tombstone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionTombstone"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Not found or not owned
        "409":
          description: Version conflict / already deleted / locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionVersionConflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived admin JWT (AP-085). iss/aud env-isolated; includes actor/role/sid.
    adminBreakGlass:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: |
        Break-glass only. Requires X-Admin-Actor + X-Admin-Reason.
        Production rejects shared key as regular auth unless ADMIN_BREAK_GLASS_ENABLED=true.
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Deprecated alias for break-glass; prefer adminBearer.
...
  headers:
    RequestID:
      description: Request correlation ID
      schema:
        type: string
    RetryAfter:
      description: Seconds to wait before retry
      schema:
        type: integer
  responses:
    BadRequest:
      description: Bad request
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServiceUnavailable:
      description: Dependency unavailable
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceUnavailableError"
...
  schemas:
...
    Error:
      type: object
      required: [error, reason_code, retryable]
      description: |
        Unified API error envelope. All 4xx/5xx JSON errors should converge here.
        Handlers may attach extra diagnostic fields, but clients must rely on these core keys.
      properties:
        error:
          type: string
          description: Human-readable error message (safe for clients)
          example: invalid JSON body
        reason_code:
          type: string
          description: Machine-readable reason code
          example: bad_request
        request_id:
          type: string
          description: Correlation id (also returned as X-Request-ID header)
        retryable:
          type: boolean
          description: Whether the client may safely retry the request
          example: false
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (e.g. unknown field name)
          example:
            field: extra
...
    ServiceUnavailableError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required: [error, reason_code, retryable]
          properties:
            error:
              type: string
              example: service unavailable
            reason_code:
              type: string
              example: db_unavailable
            retryable:
              type: boolean
              example: true
...
    CollectionAnimalDetail:
      type: object
      description: |
        Full detail for an active animal, or a minimal tombstone when deleted.
        Tombstone fields only: uuid, deleted_at, server_version.
      properties:
        uuid: { type: string, format: uuid }
        device_id: { type: string }
        account_id: { type: string }
        species: { type: string }
        species_label_zh: { type: string }
        breed: { type: string }
        rarity: { type: integer }
        hp: { type: integer }
        atk: { type: integer }
        def: { type: integer }
        spd: { type: integer }
        class: { type: string }
        element: { type: string }
        city: { type: string }
        geohash: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        generated_at: { type: string, format: date-time }
        inference_request_id: { type: string }
        nickname: { type: string, maxLength: 32 }
        favorite: { type: boolean }
        locked: { type: boolean }
        server_version: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time }
    CollectionAnimalPatch:
      type: object
      description: Partial update; at least one of nickname/favorite/locked required.
      properties:
        nickname: { type: string, maxLength: 32 }
        favorite: { type: boolean }
        locked: { type: boolean }
        server_version:
          type: integer
          format: int64
          description: Expected version when If-Match header is absent
    CollectionTombstone:
      type: object
      required: [uuid, server_version]
      properties:
        uuid: { type: string, format: uuid }
        deleted_at: { type: string, format: date-time }
        server_version: { type: integer, format: int64 }
    CollectionVersionConflict:
      type: object
      required: [error, reason_code]
      properties:
        error: { type: string }
        reason_code:
          type: string
          enum: [version_conflict, already_deleted, animal_locked]
        current:
          description: Current server representation for client merge
          oneOf:
            - $ref: "#/components/schemas/CollectionAnimalDetail"
            - $ref: "#/components/schemas/CollectionTombstone"
...